sketch: when training save different sketch file at each epoch
I like saving a generating an image each epoch, but I also like being
able to review past cycles to see visually if the results are
changing or the model seems to be converging. I've found with
this handwriting RNN in particular, the results from epoch
to epoch can be surprising, and that the progression is not
always a straightforward visual progression.

That's a lot of preamble to a very simple change: a modification of
the logic of the Sample extension to save each sketch file uniquely
when training. Could easily be extended to be an option, etc.